backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
redis==5.2.1
PyJWT==2.10.1